requirements_test.txt
pytest~=5.4.2
pytest-aiohttp
pytest-asyncio
